KEEP-2 Optionally return full response details.
By default, the result only contains list of product IDs.  However, the
Keepa API response payload includes additional useful information
including the number of products that matched the set of query parameter
filters.  This information is potentially useful for the caller to know
if there were products that met the filter criteria but were not
retrieved due to the maximum number of returned products limitation
imposed by the API.